refactor(desktop): extract shell keyboard shortcuts and palette commands into hooks
Pull the global keydown listener and the desktop palette-command registration
out of DesktopShell into useDesktopShellShortcuts and useDesktopPaletteCommands
(also moving the shared getPanelToggleShortcuts helper alongside the shortcuts
hook), following the use-animated-split.ts precedent. DesktopShell now just
calls both hooks; behavior is unchanged.
