+++ apps/desktop/src/renderer/src/shell/desktop-shell.tsx
@@ -6,10 +6,9 @@ import {
   PanelStubContent,
   PanelTabContentStack,
 } from '@linkcode/ui/shell/panels';
-import type { PaletteCommand, WorkbenchShellProps } from '@linkcode/workbench';
-import { TerminalPanel, useCommandPaletteStore } from '@linkcode/workbench';
+import type { WorkbenchShellProps } from '@linkcode/workbench';
+import { TerminalPanel } from '@linkcode/workbench';
 import { Allotment, LayoutPriority } from 'allotment';
-import { noop } from 'foxact/noop';
 import { useEffect as useAbortableEffect } from 'foxact/use-abortable-effect';
 import { useLayoutEffect } from 'foxact/use-isomorphic-layout-effect';
 import { useSingleton } from 'foxact/use-singleton';
@@ -17,7 +16,6 @@ import { useCallback, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { useTranslations } from 'use-intl';
 import { useShallow } from 'zustand/react/shallow';
-import { useDesktopSettingsStore } from '../settings/store';
 import { DesktopChrome } from './chrome/chrome';
 import { DiffStatChip } from './chrome/diff-stat-chip';
 import { DESKTOP_CHROME_SPACER_CLASS } from './chrome/metrics';
@@ -38,6 +36,8 @@ import {
   SIDEBAR_MIN_SIZE,
 } from './store/model';
 import { useDesktopShellStore } from './store/store';
+import { useDesktopPaletteCommands } from './use-desktop-palette-commands';
+import { getPanelToggleShortcuts, useDesktopShellShortcuts } from './use-desktop-shell-shortcuts';
 
 export function DesktopShell({
   systemBridge,
@@ -222,116 +222,21 @@ export function DesktopShell({
     resetBottomPanelSize: resetBottomPanelLayoutSize,
   } = shellState;
 
-  // Cmd+J (Ctrl+J off-mac) toggles the bottom terminal panel, Cmd+B (Ctrl+B) the sidebar,
-  // Option+Cmd+B (Ctrl+Alt+B) the right side panel, and Cmd+K (Ctrl+K) the command palette.
-  // Captured at the window so they win even while the terminal canvas holds focus; on mac the
-  // Ctrl variants stay with the shell (Ctrl+J is a real terminal keystroke there). Matched on
-  // `code` because Option rewrites `key` on mac (⌥B yields "∫").
-  useAbortableEffect(
-    (signal) => {
-      if (desktopPlatform === null) return;
-      const isMac = desktopPlatform === 'darwin';
-      window.addEventListener(
-        'keydown',
-        (event) => {
-          // `inert` on the hidden workbench doesn't stop window-level listeners, so Settings
-          // being open must be checked here at event time — not in the effect deps, or the
-          // listener would re-register on every open/close.
-          if (useDesktopSettingsStore.getState().settingsOpen) return;
-          const modifier = isMac
-            ? event.metaKey && !event.ctrlKey
-            : event.ctrlKey && !event.metaKey;
-          if (!modifier || event.shiftKey) return;
-          const toggle =
-            event.code === 'KeyJ' && !event.altKey
-              ? () => togglePanel('bottom')
-              : event.code === 'KeyB'
-                ? event.altKey
-                  ? () => togglePanel('right')
-                  : () => updateSidebarOpen((open) => !open)
-                : event.code === 'KeyK' && !event.altKey
-                  ? () => useCommandPaletteStore.getState().toggle()
-                  : null;
-          if (toggle === null) return;
-          event.preventDefault();
-          event.stopPropagation();
-          toggle();
-        },
-        { capture: true, signal },
-      );
-    },
-    [desktopPlatform, togglePanel, updateSidebarOpen],
-  );
+  useDesktopShellShortcuts({ desktopPlatform, togglePanel, updateSidebarOpen });
 
   const pickDirectory = useCallback(
     () => systemBridge.fs.pickFile({ title: 'Choose working folder', directory: true }),
     [systemBridge],
   );
 
-  const tPalette = useTranslations('workbench.palette');
-  // Desktop-owned palette entries: native folder pick, the Settings overlay, and the panel
-  // toggles. Registered into the shared palette store so the workbench-rendered palette lists
-  // them without desktop threading props through the surface.
-  useAbortableEffect(() => {
-    const shortcuts = getPanelToggleShortcuts(desktopPlatform);
-    const commands: PaletteCommand[] = [
-      {
-        id: 'desktop.open-folder',
-        label: tPalette('openFolder'),
-        keywords: ['folder', 'workspace'],
-        run() {
-          // The native picker only yields existing directories, so registration failing is a
-          // transport-level problem already surfaced by the data layer's error reporting.
-          void pickDirectory()
-            .then((picked) => (picked ? onRegisterWorkspace(picked) : null))
-            .catch(noop);
-        },
-      },
-      {
-        id: 'desktop.toggle-sidebar',
-        label: tPalette('toggleSidebar'),
-        shortcut: shortcuts.sidebar,
-        run() {
-          updateSidebarOpen((open) => !open);
-        },
-      },
-      {
-        id: 'desktop.toggle-bottom-panel',
-        label: tPalette('toggleBottomPanel'),
-        shortcut: shortcuts.bottom,
-        run() {
-          togglePanel('bottom');
-        },
-      },
-      {
-        id: 'desktop.toggle-right-panel',
-        label: tPalette('toggleRightPanel'),
-        shortcut: shortcuts.right,
-        run() {
-          togglePanel('right');
-        },
-      },
-    ];
-    if (onOpenSettings) {
-      commands.splice(1, 0, {
-        id: 'desktop.settings',
-        label: tPalette('openSettings'),
-        shortcut: shortcuts.settings,
-        run: onOpenSettings,
-      });
-    }
-    const { registerCommands, unregisterCommands } = useCommandPaletteStore.getState();
-    registerCommands('desktop', commands);
-    return () => unregisterCommands('desktop');
-  }, [
+  useDesktopPaletteCommands({
     desktopPlatform,
-    tPalette,
     pickDirectory,
     onRegisterWorkspace,
     onOpenSettings,
     togglePanel,
     updateSidebarOpen,
-  ]);
+  });
 
   function resetSidebarSize(): void {
     sidebarSplit.setPaneSize(DEFAULT_LAYOUT.sidebarW);
@@ -616,24 +521,4 @@ function createPanelContentHost(): HTMLDivElement {
   const host = document.createElement('div');
   host.className = 'absolute inset-0';
   return host;
-}
-
-function getPanelToggleShortcuts(platform: NodeJS.Platform | null): {
-  sidebar?: string;
-  bottom?: string;
-  right?: string;
-  palette?: string;
-  settings?: string;
-} {
-  if (platform === null) return {};
-  if (platform === 'darwin') {
-    return { sidebar: '⌘B', bottom: '⌘J', right: '⌥⌘B', palette: '⌘K', settings: '⌘,' };
-  }
-  return {
-    sidebar: 'Ctrl+B',
-    bottom: 'Ctrl+J',
-    right: 'Ctrl+Alt+B',
-    palette: 'Ctrl+K',
-    settings: 'Ctrl+,',
-  };
 }
